requirements.txt
# requirements.txt  (MySQL edition)
Django==4.2.*
djangorestframework
mysqlclient         # C‑based driver, fastest
python-dotenv       # for .env parsing
